Resolved the Some Changes status:Successfully

+++ app/src/main/java/com/example/mpos/data/occupied/json/ItemMaster.kt
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName
 data class ItemMaster(
     @SerializedName("Barcode") val barcode: String,
     @SerializedName("foodAmt") val foodAmt: Double,
-    @SerializedName("foodQty") val foodQty: Int,
+    @SerializedName("foodQty") val foodQty: Double,
     @SerializedName("id") val id: Int,
     @SerializedName("ItemCategory") val itemCategory: String,
     @SerializedName("ItemCode") val itemCode: String,
